refactor: use new CidHashMap implementation
