+++ src/db/car/plain.rs
@@ -60,7 +60,7 @@
 //! - CARv2 support
 //! - A wrapper that abstracts over car formats for reading.
 
-use crate::ipld::{CidHashMap, CidHashMapEntry};
+use crate::cid_collections::{hash_map::Entry as CidHashMapEntry, CidHashMap};
 use crate::{
     blocks::{Tipset, TipsetKeys},
     utils::encoding::from_slice_with_fallback,
@@ -196,10 +196,10 @@ where
 {
     #[tracing::instrument(level = "trace", skip(self))]
     fn get(&self, k: &Cid) -> anyhow::Result<Option<Vec<u8>>> {
-        match (self.index.read().get(*k), self.write_cache.read().get(*k)) {
+        match (self.index.read().get(k), self.write_cache.read().get(k)) {
             (Some(_location), Some(_cached)) => {
                 trace!("evicting from write cache");
-                Ok(self.write_cache.write().remove(*k))
+                Ok(self.write_cache.write().remove(k))
             }
             (Some(UncompressedBlockDataLocation { offset, length }), None) => {
                 trace!("fetching from disk");
@@ -258,7 +258,7 @@ fn handle_write_cache(
     k: &Cid,
     block: &[u8],
 ) -> anyhow::Result<()> {
-    match (index.get(*k), write_cache.entry(*k)) {
+    match (index.get(k), write_cache.entry(*k)) {
         (None, Occupied(already)) => match already.get() == block {
             true => {
                 trace!("already in cache");
